pyproject.toml
[project]
name = "stock-insight-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "numpy<2.0",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "protobuf==3.20.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "spacy>=3.8.7",
    "streamlit>=1.50.0",
    "yfinance>=0.2.66",
]
